# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Tracker API",
    "version": "1.0.0",
    "description": "REST API for the Job Tracker application.\n\n## Authentication\n\nTwo auth methods are supported:\n\n- **BearerAuth** – `Authorization: Bearer <token>`. Per-user API token generated in the dashboard. Scoped to the token owner's data (admin tokens get cross-tenant read access).\n- **CookieAuth** – Session cookie set by the Better Auth `/api/auth` flow (Google OAuth).\n\nToken management (POST/DELETE `/api/token`) requires a session cookie. The file download endpoint also accepts a `PUBLIC_READ_TOKEN` query parameter for share links."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-user API token generated via the dashboard (POST /api/token). Scoped to the token owner's data. Admin users' tokens get cross-tenant read access."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie issued by the Better Auth flow at `/api/auth`. Admin users (isAdmin=true) bypass per-user filtering."
      }
    },
    "schemas": {
...
      "ApiTokenInfo": {
        "type": "object",
        "description": "API token metadata (never includes the raw token).",
        "required": ["id", "name", "createdAt", "lastUsedAt"],
        "properties": {
          "id":         { "type": "string",          "example": "1" },
          "name":       { "type": "string",          "example": "default" },
          "createdAt":  { "type": "string",          "format": "date-time" },
          "lastUsedAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "Unauthorized" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean", "example": true }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/token": {
      "get": {
        "operationId": "getApiToken",
        "summary": "Get current API token info",
        "description": "Returns metadata about the authenticated user's API token, or null if none exists. Does not return the raw token.",
        "tags": ["Token"],
        "security": [
          { "CookieAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Token metadata or null.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "oneOf": [
                        { "$ref": "#/components/schemas/ApiTokenInfo" },
                        { "type": "null" }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "createApiToken",
        "summary": "Generate a new API token",
        "description": "Generates a new per-user API token. If a token already exists, it is revoked first. The raw token is returned only once in the response. **Requires a session** (Bearer tokens cannot create new tokens).",
        "tags": ["Token"],
        "security": [
          { "CookieAuth": [] }
        ],
        "responses": {
          "201": {
            "description": "Newly created token. The `raw` field contains the token value (shown only once).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["raw", "token"],
                  "properties": {
                    "raw":   { "type": "string", "example": "jt_abc123...", "description": "The raw API token. Copy it now — it will not be shown again." },
                    "token": { "$ref": "#/components/schemas/ApiTokenInfo" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "operationId": "revokeApiToken",
        "summary": "Revoke API token",
        "description": "Revokes the authenticated user's API token. **Requires a session.**",
        "tags": ["Token"],
        "security": [
          { "CookieAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Token revoked.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
...
}